Add using of connectionString in FilmDbContext
While running the app, the was an error: "Entity Framework provider found for the ADO.NET provider with invariant name 'System.Data.SqlClient'". This library was installed in Cinema.Data.EntityFramework, but not in CinemaPresentation.Wpf. To fix this problem we need to copy this dll every time to /debug folder or generate it every time the FilmDbContext is created.